requirements.txt
python-telegram-bot[aiohttp]==20.7
aiohttp==3.9.3
requests==2.31.0
